fix(orc): cycle collection causing heap corruption (#1106)
## Summary

In case there was a reference cycle that was kept alive by a non-
traced cell, and the non-traced cell was kept alive by a garbage
reference cycle, running the cycle collector made the collector
recurse.

A non-traced cell refers to a managed heap cell that's not scanned by
the cycle collector because its not cyclic (including due to the use
of `.acyclic`).

Depending on the circumstances, the recursion would either cause a
stack overflow or a heap corruption. The bug is now fixed.

## Details

The fundamental problem is that a call to `free` can add new roots to
the global `roots` list, either via automatic destructor or through
user-defined destructors.

Since neither the `roots` list nor root threshold are reset prior to
the cleanup phase, a call to `registerCycle` would trigger a
`collectCycles` recursion, which then ran trial deletion for all
already scanned roots again.

This would then cause a heap corruption if one of the roots was freed
already.

### The solution

Prior to the cleanup phase, all processed roots are removed from the
`roots` list. In addition, a reentrance guard (`collectorLock`) is
added that prevents recursion into `collectCycles`/`partialCollect`,
even if explicitly requested via `GC_fullCollect`.

So that a single full collection frees *all* reference cycles, even
those only discovered during a run of the cleanup phase,
`collectCycles` runs the core cycle collection in a loop until no more
new roots were discovered during cleanup.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>