tests: run more tests with all targets by default
Summary
=======
Enabled all test targets for some tests. Better test coverage for JS and
VM.

Details
=======
Reordered the target category strings to be in alphabetical order to make
later adjustments easier.
The 'targets' key is removed for all tests.
'write(stdout, ...)' is replaced with 'echo ...' to support the JS and
VM targets.
Some 'static: ' blocks have been removed and replaced with the VM target
where not specific to compile-time execution.
Tests that fail but should work are marked as 'knownIssue'.