compiler: enable ORC when running the leak test
Summary
=======

Even when ORC is disabled at run-time, cells potentially part of
reference cycles are added to potential cycle-root list, preventing
the leak sanitizer from noticing issues with un-freed cycles.

The `leakTest` define is now passed to the compiler when leak-
testing, with ORC enabled when the define is present.