sem: clarify macro pragma with bad inputs error (#1124)
## Summary

Typed macro pragmas now return a clear error message when provided bad
input AST (i.e.  `nkError` ).

## Details

The majority of changes were in  `sigmatch.matchesAux` , it now provides
the precise argument that was used for matching against the formal
parameter. In addition, when determine if there is an error at the
current parameter position it takes into account whether the formal
parameter allows  `nkError`  (i.e. is  `typed` ). This also resulted in
some minor changes, largely improvements, to error messages reported, so
the impacted test was updated ( `tsigmatch` ).

The primary impact is now when error AST is passed to a  `typed`  macro,
the programmer no longer receives an error stating that the the pragma
is invalid, instead the error from the input AST is reported. This is
true for  `typed`  and  `varargs[typed]`  parameters.

Additional fixes that were made as part of this:

-  `sigmatch.paramTypesMatchAux`  was updated to apply  `typeRel`  and
match index score changes when receiving an error node for a 
`typed/varargs[typed]`  formal parameter
-  `evaltempl.evalTemplateArgs`  wasn't scanning all args, the index in
the loop was hardcoded to  `1` , instead of  `i` , a typo
-  `evaltempl.evalTemplateArgs`  did not recursively scan args for 
`nkError` , this meant that  `varargs[typed]`  were ignored, this was
replaced with the  `walkErrors`  iterator
-  `errorhandling.buildErrorList`  couldn't handle  `nil`  nodes in its
traversal, this led to NPE when scanning for errors in template args

Fixes https://github.com/nim-works/nimskull/issues/1108

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>