niminst: use 7-Zip to create zip archives
While we originally moved away a while back, there are a few issues with Info-ZIP:

- It is rather hard to handle on Windows with the current stdlib
- It is not as widely available on Windows systems as 7-Zip

The original concerns that leads to moving away from this was that Info-ZIP has explicit
support for omitting some metadata. While revisiting 7-Zip I noticed that it appears that
7-Zip doesn't support these extra metadata in the first place. So let's give this a go
and see if the archives are reproducible.