niminst: use 7-Zip to create zip archives  (#1128)
## Summary

This pull switches from using Info-ZIP to 7-Zip for creating Zip
archives. This makes sure that the build works correctly on Windows
regardless of locale.

## Details

While we originally moved away a while back, there are a few issues with
Info-ZIP:

- It uses "legacy" APIs on Windows, and is not Unicode aware on such
systems. This makes it unable to ingest the file list on Windows
depending on locale.
- It is not as widely available on Windows systems as 7-Zip
- Development of Info-ZIP has largely stopped, so the issues we found
are unlikely to be fixed.

The original concerns that lead to moving away from this was that
Info-ZIP has explicit support for omitting some metadata. Reexamining
7-Zip, I found that:

- 7-Zip does not support these extra metadata in the ancient version
used by most Linux distros. Later versions of 7-Zip supports these, but
have them disabled by default.
- While there are some problems with filename encoding if locale does
not support UTF-8, this problem seems to be isolated to Unix and can be
solved by forcing the locale to  `C.utf8` , which is done here.