shellrunner: add Windows support (#1129)
## Summary
This pull request adds support for Windows to `std/shellrunner`.

## Details
Most of the code is already there, having originated from osproc. Only a
new API has to be added to osproc to expose the process handle for
Windows.

In the future the process handle API could be expanded to POSIX systems
where pidfd is available (ie. Linux, FreeBSD).