vmgen: fix `mExpandToAst` calls
The template callee node needs to reach `vmgen` as-is, but this is no
longer possible with `mnkProc` not storing a `PSym` (and the template
symbol was also never a normal procedure anyway).

`mirgen` now turns the callee symbol into an AST value (`mnkLiteral`),
so that `vmgen` can retrieve the `PNode`.