prevent incremental procedures from being processed normally
Incremental procedures (i.e., those whose content is discovered
incrementally) must not be queued for normal processing.

In order to prevent this, all possible incremental procedures are
marked with `sfForward` prior to the processing loop, with `queue`
then ignoring all procedures marked with the flag.

Once all alive procedures were processed, all incremental procedures
that now have a body have the `sfForward` flag excluded again. This is
to prevent interfering with the code generator, which might still
expect this flag to not be present.