fix(mirgen): loop-initialized closure locals not being destroyed (#1134)
## Summary

Fix locals defined within a loop and closed over by an inner procedure
(which lifts them into the closure environment) not properly destroyed
when control-flow reaches the `let` or `var` statement again.

The compiler crashing for `let (a, b) = x` statements where `a`
or `b` are closed over is fixed too.

## Details

Except for `let :env.a = b` statements, `mirgen` treated all `let`/
`var` statements for closure locals as *initializing* assignments (by
emitting an `mnkInit` statement).

This is correct for closure locals defined outside of loops, but not
for those defined within loops, as control-flow possibly reaches the
assignment multiple times, but location having been destroyed in
between, (the closure environment is the same through the whole
procedure invocation).

Since `mnkInit` assignments are (correctly) turned into raw
assignments, the old value wasn't destroyed on re-assignment. `mirgen`
now keeps track of whether a statement is enclosed by a loop (via the
`inLoop` counter), and if the statement is, emits a normal assignment
(`mnkAsgn`) instead of an initializing one.

### Compiler crash

For `let (a, b) = x`, `mirgen.genVarTuple` always tried to access the
identifiers nodes as if they were symbols. If they're closed over, this
results in a field defect, since then they're `nkDotExpr` nodes. The
type is now taken from the node itself.

The added test also covers this.