sem: `auto`/inferred lambdas generate valid AST (#1135)
## Summary
`auto` parameter and return types resulted in inferred lambda AST being
generated with `nil`s in the parameter position. This is now no longer
the case and the AST generated is well formed.

## Details

`seminferredlambda` was incorrectly copying the AST out of the `tyProc`
type into the parameter position; that AST isn't well formed. This
resulted in `nil` nodes being inserted for the parameter and return in
the output AST. When traversing it for errors, this would generate a
NPE
during `buildErrorList`. A `nil` checking guard was put in place in
order to work around the problem.
see: https://github.com/nim-works/nimskull/pull/1124

Now the AST is well formed, and the check has been removed.
Additionally, some associated changes have been made:

- tracing coverage expansion to diagnose the issue and confirm the fix
- noting potentially the same issue in `liftdestructors`

---------

Co-authored-by: zerbina <100542850+zerbina@users.noreply.github.com>