fix(vmbackend): run-time errors with non-zero-based sets
Summary
=======

Fix a bug with the VM backend's constant data serialization where
having and using a constant `set` with a non-zero lower bound resulted
in:
* set operations on the set not working as expected
* runner crashes at startup

Details
=======

The deserialization logic of `vmrunner` expects zero-based `set`
construction elements, but the serialization logic in `packed_env`
didn't make sure this is the case.

`PackedEncoder` now requires access to a `ConfigRef`, which is used for
subtracting the lower bound from all element values (making them zero
based).