fix(vmbackend): run-time errors with non-zero-based sets (#1138)
## Summary

Fix a bug with the VM backend's constant data serialization where
having and using a constant `set` with a non-zero lower bound resulted
in:
* set operations on the set not working as expected
* runner crashes at startup

## Details

The de-serialization logic in `vmrunner` expects zero-based `set`
construction elements, but the serialization logic in `packed_env`
didn't ensure that the element values are zero-based.

`PackedEncoder` now requires access to a `ConfigRef`, which is used for
subtracting the lower bound from all element values (making them zero
based).