vmgen: serialize constant *data* instead of constants
The difference is somewhat subtle at the moment: constant data is
the actual data, whereas constants associate a name/interface with
constant data.

As a consequence, `DataId` replaces `ConstId` for the link index.

The difference between user- and compiler-defined constants doesn't
matter to the VM -- the VM only cares about the underlying data.