mirtrees: support anonymous constants
`ConstId` is changed to represent both user-defined and anonymous
constants. The most-significant bit indicates which of the two a
`ConstId` represents: if the bit is set, the lower 31 bits represent
a `DataId`, otherwise the `ConstId` identifies a user-defined
constant.

This encoding is backwards-compatible (>= 2^31 user-defined constants
is highly unlikely if not impossible).