mirgen: lift constant expressions
This is essentially the same as what `cgen` currently does, but with
the difference that `mirgen` can - and does - disable the lifting when
an expression is used in a sink context.

Not lifting an expression used in a sink context is preferred, as
constants have to be copied, whereas the raw construction can
construct the value in place (no copy).

As a temporary workaround for `DataTable` not support hidden
conversions in constant expressions, a custom implementation of
`isDeepConstExpr` is used.