jscore: remove `copyWith` (#1141)
## Summary

The binding to the JavaScript `copyWith` function is implemented
incorrectly (it's a mutating function, but the parameter type is
`openArray`, which is immutable), which means that
`copyWith([...], ...)` relies on undefined behaviour.

Since there are also no more elaborate tests beyond the
`runnableExamples`, the procedure is removed for now.

## Details

If re-added in the future, the binding should use `var openArray` (and
return nothing) instead of `openArray`.