fix(vm): access check not failing (#1142)
## Summary

Fix a bug with the VM's pointer-to-cell mapping that allowed out-of-
bounds reads or writes of empty `seq`s to crash the VM, instead of
resulting in a proper memory access error.

## Details

The fundamental problem was with unclear responsibilities: callers of
`mapPointerToCell` and `mapInteriorPointerToCell` expected both to
handle `nil` pointers, while the procedures expected their callers to
do so.

At the moment, the `mapX` procedure consider free cells (for which
`VmCell.p == nil`), which means that trying to map `nil` pointer values
could result in a valid `CellId` being returned! A `LocHandle` having a
non-`-1` `CellId` means that only the cell's data pointer (`nil`, in
this case) and the user pointer need to match (which in this specific
case, they did) for the access checks to not reject the access.

The dangerous `LocHandle` could only be created through indexing
into a `seq`, as the `makeLocHandle` overload for arbitrary user-
pointers properly guarded against `nil` pointers. If the VM was built
with assertions enabled, merely assigning such handle to a register
(incorrectly) resulted in an assertion failure.

When the VM is built with assertions disabled, reading from or writing
to the location resulted in the VM crashing with a segmentation
fault / `nil` dereference error.

The incorrect assertion (forbidding `nil` handles in registers is
wrong) is removed, and `mapPointerToCell`/ `mapInteriorPointerToCell`
now always return a lookup failure for `nil` pointers, fixing the
two problems.

Two tests for situations where a crash would previously happen are
added.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>