fix(cgen): run-time heap corruption with static `seq` (#1120)
## Summary

Fix a C code generator bug where using the result of a `static`-
evaluated expression (i.e., `static: x`), resulted in a heap corruption
at run-time. Only values that had a generic `seq` alias as their type
were affected.

Fixes https://github.com/nim-works/nimskull/issues/1117.

## Details

The problem is that `genSeqConstr` didn't call `skipTypes` on the
`cnkArrayConstr` (a `seq` construction, in this case) expression. This
works if the type is a `tySeq`, but for `tyGenericInst`, `t.lastSon`
returns the `tySeq`, and treating the `tySeq` as the element is wrong.

Since the size and alignment of the `seq` type itself are then passed
to `newSeqPayload`, to little space is allocated, causing an out-of-
bounds memory access if the actual elements' size is larger than
`sizeof(seq[T])`.

`genSeqConstr` now properly skips the type, fixing the bug.
`genArrToSeq` was also missing the type skipping, it's adjusted too.