transf: elide `else`-branch-only `case` statements
The case statement/expression is unnecessary if it only has an else
branch; these case statements are now transformed into just a
statement-list(-expression).

To make sure selector expression side-effects (if any) are still
computed, the expression is transformed into a
`discard selector` statement.

For making sure lifetime scopes don't change, the body of the `else` is
wrapped in a `block`.