remove `posix` dependency from `timportedobj` (#1149)
## Summary

The reliance on the `posix` module made test not usable on platforms
not supporting the module. The test is reduced in order to not depend
on `posix`.

## Details

The test was added as a regression test for a bug with
`liftdestructors`  (fixed by

https://github.com/nim-works/nimskull/commit/c87796180edc86cf3cc1fa0fb35f03c1461937cd),
where a hook was lifted for imported types. If the imported type
contains a field that doesn't actually exist in the imported type (this
is the case for, e.g., padding), the resulting assignment of those
fields would result in a C compiler error.

The test is changed in a way such that it's as explicit as possible (no
`seq` usage, no reliance on `Stat`).