io, os: use 1-parameter newWideCString to prealloc
The 2-parameter form of newWideCString is meant to take a cstring and
its length, then convert them to wide form. As such this length must be
the exact length of the given string.

This commit fixes buffer overflow due to wrong usage of this API, where
an empty string was given in the context of preallocation by switching
to the 1-parameter form which is meant for that.