fenv, math: don't specify -lm on macOS
Xcode 15, which is used by newer macOS now warns if the flags are redundant,
breaking tests. Suppressing such warnings does not seem to be feasible, as it
was newly added and suppression flags would not work on older Xcode.

However, it appears that macOS doesn't have libm, only a stub for POSIX compliance,
so we could just omit the flag altogether.