sem: fix infinite loop in type identifier lookup
Summary
=======

Fixed an infinite loop in type identifier lookup, when parameter names
collided with parameter type declarations.

fixes: https://github.com/nim-works/nimskull/issues/1151

Details
=====

`semtypes.semTypeIdent` had an infinite loop when searching for
potentially ambiguous symbols. This meant code with ambiguities between,
for instance, a parameter and a type name resulted in an infinite loop.

In addition to the change a test has been added to avoid the regression.