compilerbridge: use `initFromExpr` for `putIntoReg`
Instead of using the `PNode`-based initialization, the expression is
first translated to the MIR, and then loaded into the location with
`initFromExpr`.

The new `constDataToMir` procedure provided by `vmjit` delegates to
the normal AST -> MIR node translation, while also taking care of the
registering procedures discovered during the translation with the VM
environment.

This `vmaux.lookupProc` procedure is obsolete and thus removed.