ci: bump x86 macOS image to version 12
Following macOS 14 support announcement, GitHub has officially
deprecated version 11 runners. As such, migrate to the next version to
prevent disruptions when GitHub decides to retire these runners.

This meant that our binaries will target macOS 12 by default, which is
around 3 years old. I expect the binaries to still work on older
versions, but we cannot guarantee that to the future.

Refs https://github.blog/changelog/2024-01-30-github-actions-macos-14-sonoma-is-now-available/