vm: remove string constants (#1158)
## Summary

Remove support for string constants from the VM -- strings have to be
treated as data and loaded as such now. This slightly simplifies the VM
and speeds up loading string literals, as no allocation + string copy
is necessary when loading a string literal for non-mutating use (e.g.,
`f("abc")`).

## Details

String constants in the VM had two problems:
- for the VM guest to work with strings, the string has to be a VM
  memory location. Since VM constants are not, this required the
  allocation of a memory location plus a string copy
- to prevent duplicate string constants, a linear search over all
  string constants was performed, which is very inefficient (especially
  if there are many large string constants)

In order to resolve both problems and simplify the VM, a MIR pass is
used to lift anonymous MIR constants from all string literals and
replace the string literals with the anonymous constant. In effect,
this means that no `cnkStrLit` node reaches `vmgen`, and that all
string literals are treated as normal data.

String VM constants (`cnstString`) and the `AsgnConst` VM opcode (which
was only used for assigning string constants) are thus obsolete and
removed.

For the lifting pass to work, the following changes are required:
- a mutable `MirEnv` has to be passed to `mirpasses.applyPasses`
- the MIR has to allow `mnkConst` nodes as branch labels
- all string literals reaching the MIR must be typed. All problematic
  `newStrNode` usages are changed to also pass a type

The VM speed up is the result of loading a string literal via
`LdCmplxConst`, which only loads a handle and doesn't perform a copy
itself.