std/io, os: stop `newWideCString` buffer overflow (#1153)
## Summary

Fix buffer overflow in the  `std/io`  and  `std/os`  modules related to
incorrect usage of  `newWideCstring` .

## Details

`newWideCString` 's binary form, 
`proc newWideCString*(source: cstring, L: int): WideCStringObj` ,
results in buffer overflows within the  `std/io`  and  `std/os` 
modules. This change revises the erroneous usage of  `newWideCString` ,
with an empty string ( `""` ) as the  `source`  parameter in these
modules, with the single parameter version for preallocating a wide
string.

The 2-parameter form of newWideCString is meant to take a  `cstring` 
and its  `length` , then convert them to wide form. As such this length
must be the exact length of the given string.

To prevent incorrect usage in the future, this PR also documents the
various  `newWideCString`  variants.