vmgen: clean up some linking-related types (#1163)
## Summary

Remove the obsolete `symToIndexTbl` table and the `LinkerData` type,
and rename `vmlinker` to `identpatterns`.

## Details

The `symToIndexTbl` table was originally used for mapping a `PSym` to
the entity's link index, but the `MirEnv` with its `SymbolTable` fully
subsumed this -- the remaining usages of `symToIndexTbl` are replaced
with querying `MirEnv` instead.

As a consequence, some routine interfaces change:
* `vmjit.isAvailable` has an additional `JitState` parameter
* `setGlobalValue` and `getGlobalValue` require access to the whole
  `EvalContext`
* `vmjit` provides the new `getGlobal` for querying the link index for
  the symbol of a global

Furthermore, since there's no need for a separate `LinkerData` type
anymore, all `LinkerData`  usage is replaced with directly embedding/
using the identifier pattern list. `LinkIndex` is moved to `vmdef` and
the `vmlinker` module, which contains the identifier pattern matching,
is renamed to `identpatterns`.