tests: update `toverflow_abs`
The test succeeds with the C target now, but still fails with the JS
backend, since the negation doesn't fail properly.

In addition, the extra invocation to "instantiate" `abs` is removed:
`abs` for integers is not a generic routine, so the call is
unnecessary.

Finally, the expected output is changed such that is also matches for
the C target.