mirgen: make sure `mAbsI` uses a checked call
`mAbsI` is not in the set of magics that can raise an exception,
meaning that `genCall` would not emit a checked call for it, thus
hiding the exceptional control-flow.

When panics are disabled and the used `abs` magic has overflow checks
enabled, `mirgen` now uses a dedicated path for making sure that a
`mnkCheckedCall` is used.