fix(vmgen): wrong `range` value being read (#1167)
## Summary

Fix an incorrect/invalid value being read from locations of unsigned
`range` type. Only the VM backend was affected, and only when the value
stored in the location had the most-significant bit set.

## Details

With the VM, all unsigned values require masking after being read from
a memory location.  `genRegLoad`  emits the necessary masking
instruction,
but `isUnsigned` doesn't consider range types (a separate bug), meaning
that values of type `tyRange` with an unsigned base type were not
masked.

* all code generation for narrowing/masking instructions is changed to
  skip `tyRange` types
* the custom narrowing implementation used for `mShrI` code generation
  is merged into `genNarrow`
* the duplicated unsigned integer masking is moved to a dedicated
  procedure (`genNarrowUnsigned`)

There was also the problem of the types' `size` field being accessed
directly; since a type's size is computed on demand, the value in
`size` might not be valid. `getSize` is used instead.