vmgen: implement the float variants of built-in arithmetic ops
The same bytecode is generated for the built-in float ops as is for the
float arithmetic magics. Nothing changes for the magics, as float
overflow checks aren't yet implemented for the VM.

As for the implementation, the existing `cnkNeg`, `cnkAdd`, etc. code
generation is re-used, but the narrowing is removed (it's not valid for
floats).

Not using the narrow instructions also speeds up the unchecked signed
integer arithmetic.