cgen: implement the float variant of built-in arithmetic ops
They're translated in the same way as unchecked float arithmetic has
been translated so far. For convenience, the `compat.pick` procedure is
added, for selecting one of two enum values based on an expression's
type.

The `binaryFloatArithmetic` procedure, which is invoked for
translating magics for float arithmetic, unconditionally emits the
infinity checks.