sem: correctly show errors from `defer` node body
Summary
=======

Errors in a defer node body are now correctly reported, no longer
producing misleading `'defer' takes a 'void' expression` messages.

Details
======

A`defer` with an error node body is no longer treated as an
expression and having the underlying error supressed. Instead the body
error is correctly reported. The `defer` AST is also correctly wrapped
in an error node.

A relatively comprehensive test sutie for `defer` statements was added,
along with an improvement to the void expression error message
requirement, which shows the `defer` body itself.